feat: add fusion_2ch/3ch templates, .gitignore, README, and training scripts
- Support fusion_2ch and fusion_3ch template types in finetune.py and inference.py
- Add .gitignore to exclude large data/model files
- Rewrite README in Chinese with full usage guide
- Add train_lightgcn.py for collaborative embedding training
- Add data_preprocess/ml-1m.py preprocessing script
- Update RAG modules and prompt templates

+++ rag/get_semantic_embed_rich.py
@@ -14,7 +14,7 @@ embeds = []
 
 def hook(module, input, output):
     global cur_embed, embeds
-    input = input[0].cpu().detach().numpy()
+    input = input[0].cpu().detach().float().numpy()
     cur_embed = input
